requirements.txt
requests>=2.31
openai>=1.0
pyyaml>=6.0
python-dotenv>=1.0
supabase>=2.0
psycopg2-binary>=2.9
